Test how servos work
Did some light testing with continuous and stepper servos we found in the lab. Specifically, Feetech FT90R continuous servos, and a Sunfounder 55g stepper servo.
The continuous servo is better, and seems to range from wanting 2% duty cycle to 12% duty cycle? But switching between 7 and 12 should give us the 90 degree turn we're looking for for the flag. Yay :)